mcp-research-agent/requirements.txt
httpx==0.27.2
trafilatura==1.12.2
pdfminer.six==20231228
beautifulsoup4==4.12.3
tenacity==9.0.0
uvloop==0.20.0; platform_system != 'Windows'


